+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - 🎉 added autumn and spring themes
 - 📈 the default theme is now spring
 - 📈 improved the markdown viewer -- less clutter
+- 🐛 HTML will now be treated as markdown when no output is specified -- allows for syntax highlighted code blocks instead of just printing it back. 
 - 🐛 now removes the background color when converting HTML to image
 
 ## V0.3.6